+++ kondor-examples/src/jmh/kotlin/com/ubertob/kondor/json/jmh/DemoClassFixtures.kt
@@ -1,5 +1,6 @@
 package com.ubertob.kondor.json.jmh
 
+import com.ubertob.kondor.json.toNdJson
 import com.ubertob.kondortools.chronoAndLog
 import org.openjdk.jmh.annotations.Scope
 import org.openjdk.jmh.annotations.State
@@ -11,6 +12,7 @@ open class DemoClassFixtures {
     }.toList()
 
     val jsonString = jDemoClasses.toJson(objList)
+    val ndJsonStrings = toNdJson(JDemoClass)(objList)
 }
 
 
@@ -23,6 +25,19 @@ fun benchLoop(ser: (List<DemoClass>) -> String, deser: (String) -> List<DemoClas
             repeat(10_000) {
 //                val json = ser(testFix.objList)
                 val list = deser(testFix.jsonString)
+            }
+        }
+    }
+}
+
+fun benchLoopNd(ser: (List<DemoClass>) -> Sequence<String>, deser: (Sequence<String>) -> List<DemoClass>) {
+    val testFix = DemoClassFixtures()
+
+    repeat(1_000) {
+        chronoAndLog("iter $it") {
+            repeat(10_000) {
+                val json = ser(testFix.objList)
+                val list = deser(testFix.ndJsonStrings)
             }
         }
     }