+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 This list is not currently intended to be all-encompassing - it will document major and breaking API changes with their
 rationale when appropriate:
 
+### v.2.2.3 - 14 February 2024
+
+Kondor-core: trying to parse an empty file gives an error
+
+### v.2.2.2 - 31 January 2024
+
+Kondor-core: handling NaN and Infinity deserialization of Double
+
 ### v.2.2.1 - 10 January 2024
 
 Kondor-core: fixed issue with invalid Json parsing